+++ pybop/models/li_half_cell/__init__.py
@@ -1,0 +1,5 @@
+#
+# Import lithium half cell models
+#
+from .sp_diffusion import SPDiffusion
+from .weppner_huggins import WeppnerHuggins