+++ tests/integration/test_applications.py
@@ -48,13 +48,14 @@ class TestApplications:
         )
 
     def test_interpolant(self, parameter_values, discharge_dataset):
-        parameter_values = pybop.lithium_ion.SPDiffusion.create_grouped_parameters(
+        parameter_values = pybop.li_half_cell.SPDiffusion.create_grouped_parameters(
             parameter_values
         )
-        parameter_values["Electrode OCP [V]"] = pybop.Interpolant(
+        parameter_values["Positive electrode OCP [V]"] = pybop.Interpolant(
             discharge_dataset["Stoichiometry"], discharge_dataset["Voltage [V]"]
         )
-        model = pybop.lithium_ion.SPDiffusion(build=True)
+        parameter_values.set_initial_state(0.9)
+        model = pybop.li_half_cell.SPDiffusion(build=True)
         t_eval = np.linspace(0, 10, 100)
         solution = pybamm.Simulation(model, parameter_values=parameter_values).solve(
             t_eval=t_eval, t_interp=t_eval
@@ -186,25 +187,25 @@ class TestApplications:
     def test_gitt_pulse_fit(
         self, half_cell_model, half_cell_parameter_values, pulse_data
     ):
-        parameter_values = pybop.lithium_ion.SPDiffusion.create_grouped_parameters(
+        parameter_values = pybop.li_half_cell.SPDiffusion.create_grouped_parameters(
             half_cell_parameter_values
         )
-        diffusion_time = parameter_values["Particle diffusion time scale [s]"]
+        diffusion_time = parameter_values["Positive particle diffusion time scale [s]"]
 
         gitt_fit = pybop.GITTPulseFit(parameter_values=parameter_values)
         gitt_result = gitt_fit(gitt_pulse=pulse_data)
 
         np.testing.assert_allclose(
-            gitt_result.best_inputs["Particle diffusion time scale [s]"],
+            gitt_result.best_inputs["Positive particle diffusion time scale [s]"],
             diffusion_time,
             rtol=5e-2,
         )
 
     def test_gitt_fit(self, half_cell_model, half_cell_parameter_values, pulse_data):
-        parameter_values = pybop.lithium_ion.SPDiffusion.create_grouped_parameters(
+        parameter_values = pybop.li_half_cell.SPDiffusion.create_grouped_parameters(
             half_cell_parameter_values
         )
-        diffusion_time = parameter_values["Particle diffusion time scale [s]"]
+        diffusion_time = parameter_values["Positive particle diffusion time scale [s]"]
 
         with pytest.raises(
             ValueError, match="The initial current in the pulse dataset must be zero."
@@ -224,7 +225,7 @@ class TestApplications:
         gitt_parameter_data = gitt_fit()
 
         np.testing.assert_allclose(
-            gitt_parameter_data["Particle diffusion time scale [s]"],
+            gitt_parameter_data["Positive particle diffusion time scale [s]"],
             np.asarray([diffusion_time]),
             rtol=5e-2,
         )